Move remaining code out of eph.go
